feat: alias execute and deprecate executeFile
