+++ public/openapi.json
@@ -88,7 +88,7 @@
     "/command/execute": {
       "post": {
         "operationId": "executeCommand",
-        "summary": "Execute a command",
+        "summary": "Execute using first available mode",
         "security": [
           {
             "bearerAuth": []
@@ -199,7 +199,8 @@
     "/command/execute-file": {
       "post": {
         "operationId": "executeFile",
-        "summary": "Execute a file present on the server/target",
+        "summary": "Execute a file present on the server/target (deprecated)",
+        "deprecated": true,
         "security": [
           {
             "bearerAuth": []
