+++ public/openapi.yaml
@@ -55,7 +55,7 @@ paths:
   /command/execute:
     post:
       operationId: executeCommand
-      summary: Execute a command
+      summary: Execute using first available mode
       security:
         - bearerAuth: []
       requestBody:
@@ -122,7 +122,8 @@ paths:
   /command/execute-file:
     post:
       operationId: executeFile
-      summary: Execute a file present on the server/target
+      summary: Execute a file present on the server/target (deprecated)
+      deprecated: true
       security:
         - bearerAuth: []
       requestBody:
